+++ io.werval/io.werval.runtime/src/main/java/io/werval/runtime/ApplicationInstance.java
@@ -1087,7 +1087,11 @@ public final class ApplicationInstance
         {
             try
             {
-                list.add( (ParameterBinder<?>) classLoader.loadClass( parameterBinderClassName ).newInstance() );
+                ParameterBinder<?> binder = (ParameterBinder<?>) classLoader
+                    .loadClass( parameterBinderClassName )
+                    .newInstance();
+                binder.onActivate( this );
+                list.add( binder );
             }
             catch( ClassNotFoundException | InstantiationException | IllegalAccessException | ClassCastException ex )
             {