+++ CHANGELOG.md
@@ -18,6 +18,10 @@ For minor changes, build system or test changes, see the full [source code histo
 - #232 & #233 Fix dynamic modules documentation in development mode
 - #235 Fix QueryString.Decoder that incorrectly added a null query param
 
+### New features
+
+- #194 Hashid parameter binder
+
 ### Enhancements
 
 - #68 Server bootstrap now loads `Application` and `HttpServer` implementations through `ServiceLoader`